Move int to long

+++ src/Lykke.Service.Chainalysis.Mock.AzureRepositories/Dto/UserDepositAddressInfo.cs
@@ -21,9 +21,9 @@ namespace Lykke.Service.ChainalysisMock.AzureRepositories.Dto
             Total = Data.Count;
         }
 
-        public int Total { get; set; }
-        public int Limit { get; set; }
-        public int Offset { get; set; }
+        public long Total { get; set; }
+        public long Limit { get; set; }
+        public long Offset { get; set; }
         public List<IAddressInfo> Data { get; set; }
     }
 }
