+++ src/Lykke.Service.Chainalysis.Mock.AzureRepositories/Dto/UserTransactionInfo.cs
@@ -20,9 +20,9 @@ namespace Lykke.Service.ChainalysisMock.AzureRepositories.Dto
             Total = Data.Count;
         }
 
-        public int Total { get; set; }
-        public int Limit { get; set; }
-        public int Offset { get; set; }
+        public long Total { get; set; }
+        public long Limit { get; set; }
+        public long Offset { get; set; }
         public List<ITransactionInfo> Data { get; set; }
     }
 }
