+++ src/Lykke.Service.Chainalysis.Mock.Contracts/UserDepositAddressInfoModel.cs
@@ -4,9 +4,9 @@ namespace Lykke.Service.Chainalysis.Mock.Contracts
 {
     public class UserDepositAddressInfoModel
     {
-        public int Total { get; set; }
-        public int Limit { get; set; }
-        public int Offset { get; set; }
+        public long Total { get; set; }
+        public long Limit { get; set; }
+        public long Offset { get; set; }
         public List<AddressInfoModel> Data { get; set; }
     }
 }
