+++ src/Lykke.Service.Chainalysis.Mock.Services/Dto/UserInfo.cs
@@ -21,9 +21,9 @@ namespace Lykke.Service.ChainalysisMock.Services.Dto
             Data = userData.Cast<IUserData>().ToList();
         }
 
-        public int Total { get; set; }
-        public int Limit { get; set; }
-        public int Offset { get; set; }
+        public long Total { get; set; }
+        public long Limit { get; set; }
+        public long Offset { get; set; }
         public List<IUserData> Data { get; set; }
     }
 }
