+++ src/Lykke.Service.Chainalysis.Mock.Services/Extentions/DateTimePosix.cs
@@ -4,12 +4,12 @@ namespace Lykke.Service.ChainalysisMock.Services.Extentions
 {
     public static class DateTimePosix
     {
-        public static int ToPosix(this DateTime dateTime)
+        public static long ToPosix(this DateTime dateTime)
         {
-            return (int)(dateTime.ToUniversalTime().Subtract(new DateTime(1970, 1, 1))).TotalSeconds;
+            return (long)(dateTime.ToUniversalTime().Subtract(new DateTime(1970, 1, 1))).TotalSeconds;
         }
 
-        public static DateTime FromPosix(this int posixTime)
+        public static DateTime FromPosix(this long posixTime)
         {
             return new DateTime(1970, 1, 1).AddSeconds(posixTime).ToLocalTime();
         }
